Add a/b split
